Check block type in config finder
